requirements.txt
crewai==0.16.3
crewai-tools==0.0.12
python-dotenv==1.0.1